Build from source rather than curlpipebash
Advising users to install by piping the kr binary to bash isn't the nicest
thing in the world, but I appreciate that it makes it easily accessible to
people using a one-liner. Given this project's intentions, it seems to me
that the FAQ should prefer being secure over user convenience in situations
like this.

Update the README to recommend building from source.